refactor: add component for order book controls and fix display
